re-organizing server
