+++ dnsext-full-resolver/dnsext-full-resolver.cabal
@@ -32,8 +32,6 @@ library
         DNS.Cache.Queue
         DNS.Cache.RootServers
         DNS.Cache.RootTrustAnchors
-        DNS.Cache.ServerMonitor
-        DNS.Cache.SocketUtil
 
     default-language: Haskell2010
     ghc-options:      -Wall
@@ -66,12 +64,22 @@ library
 executable bowline
     main-is:          bowline.hs
     hs-source-dirs:   bowline
+    other-modules:
+        Monitor
+        SocketUtil
     default-language: Haskell2010
     ghc-options:      -Wall -threaded -rtsopts "-with-rtsopts=-N2 -qn1"
     build-depends:
         base,
         dnsext-log,
-        dnsext-full-resolver
+        dnsext-full-resolver,
+        dnsext-types,
+        dnsext-do53,
+        dnsext-dnssec,
+        iproute,
+        stm,
+        network,
+        unliftio
 
     if (os(windows) && impl(ghc >=9.0))
         ghc-options: -with-rtsopts=--io-manager=native
